Support extended DEM zoom level
This is the opposite of compounding raster tiles.
In this way the fetcher produces smaller high zoom level tiles that
can be textured with native raster tiles instead of having to compound
them.
Downside is the increased amount of tiles and therefore of drawcalls
that become quickly cpu bound

Also introduce basic support for saving / replaying queries